sound: sunxi: Add sun8i A33 codec and sound card drivers
Add a sunxi sun8i codec for the A33 Allwinner and create the
sun8i audio card to link the DAI and the codec driver.

This is a first implementation and some improvements are necessary.
The PRCM registers need to be handled in an aux_devices and it should
use DAPM widgets which is not the case, for example.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>